refactor(passkey): passkeyをauthenticate配下からfeatures直下へ移動
passkeyはauthenticateの内部機能ではなく独立した機能のため、
features/authenticate/passkey/ から features/passkey/ へ移す。
セッションキー(SESSION_SWR_KEY)のみ authenticate を参照する。
