+++ application/apps/web/src/client/features/passkey/use-passkey-register.ts
@@ -3,7 +3,7 @@ import type { PublicKeyCredentialCreationOptionsJSON } from '@simplewebauthn/bro
 import { wrapAsyncCall } from '@trend-diary/common/result'
 import { useState } from 'react'
 import { toast } from 'sonner'
-import { PASSKEY_MESSAGES } from '@/client/features/authenticate/passkey/model'
+import { PASSKEY_MESSAGES } from '@/client/features/passkey/model'
 import getApiClientForClient from '@/client/infrastructure/api'
 
 export default function usePasskeyRegister() {
